+++ src/Buffer.hpp
@@ -143,7 +143,7 @@ private:
     inline int32_t compute_sprite_yx_start_indx(int32_t x, int32_t y, int32_t spriteHeigh) const;
     inline int32_t compute_y_start_indx(const int32_t&);
     void append_sprite(int32_t, int32_t, const uint8_t*, int32_t, int32_t, colors::Colors color = colors::ORANGE);
-    int32_t append_digits(int32_t, int32_t, Sprite&, int32_t, size_t, colors::Colors);
+    int32_t append_digits(int32_t, int32_t, Sprite&, int32_t, int32_t, colors::Colors);
     void update_fps(void);
 };
